+++ pkg/iterator_test.go
@@ -61,7 +61,7 @@ func createMockBinaryEntry(path string) BinaryEntryInterface {
 		UID:          1000,
 		GID:          1000,
 		FileSize:     100,
-		HashType:     uint16(HashTypeSHA1),
+		HashType:     HashTypeSHA1,
 		EntryFlags:   0,
 		IsDeleted:    false,
 	}
